build: switch to golangci-lint (#2306)
* test: fix a few more linter errors

* build: switch to golangci-lint

fixes #2207

* build: use config file for linter

+++ metrics/heartbeat_test.go
@@ -183,7 +183,7 @@ func mustMakeTipset(t *testing.T, height types.Uint64) types.TipSet {
 		Ticket:          nil,
 		Parents:         types.SortedCidSet{},
 		ParentWeight:    0,
-		Height:          types.Uint64(height),
+		Height:          height,
 		Nonce:           0,
 		Messages:        nil,
 		MessageReceipts: nil,
